ODE-138 Uploader service moves files to designated directory. Improved UI. Added working Message Frame file to data/ directory

+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/storage/FileSystemStorageService.java
@@ -25,18 +25,43 @@ public class FileSystemStorageService implements StorageService {
 
     private OdeProperties properties;
     private final Path rootLocation;
+    private final Path bsmLocation;
+    private final Path messageFrameLocation;
 
     @Autowired
     public FileSystemStorageService(OdeProperties properties) {
         this.properties = properties;
 
         this.rootLocation = Paths.get(this.properties.getUploadLocationRoot());
-        logger.info("Upload location: {}", this.rootLocation);
+        this.bsmLocation = Paths.get(this.properties.getUploadLocationBsm());
+        this.messageFrameLocation = Paths.get(this.properties.getUploadLocationMessageFrame());
+
+        logger.info("Upload location (root): {}", this.rootLocation);
+        logger.info("Upload location (bsm): {}", this.bsmLocation);
+        logger.info("Upload location (message frame): {}", this.messageFrameLocation);
     }
 
     @Override
-    public void store(MultipartFile file) {
-        Path path = this.rootLocation.resolve(file.getOriginalFilename());
+    public void store(MultipartFile file, String type) {
+
+        // Discern the destination path via the file type (bsm or messageFrame)
+        Path path;
+        if (("bsm").equals(type)) {
+            path = this.bsmLocation.resolve(file.getOriginalFilename());
+        } else if (("messageFrame").equals(type)) {
+            path = this.messageFrameLocation.resolve(file.getOriginalFilename());
+        } else {
+            EventLogger.logger.info("File type unknown: {} {}", type, file.getName());
+            throw new StorageException("File type unknown: " + type + " " + file.getName());
+        }
+
+        // Check file is not empty
+        if (file.isEmpty()) {
+            EventLogger.logger.info("File is empty: {}", path);
+            throw new StorageException("File is empty: " + path);
+        }
+
+        // Check file does not already exist (if so, delete existing)
         try {
             EventLogger.logger.info("Deleting existing file: {}", path);
             Files.deleteIfExists(path);
@@ -44,17 +69,15 @@ public class FileSystemStorageService implements StorageService {
             EventLogger.logger.info("Failed to delete existing file: {} ", path);
             throw new StorageException("Failed to delete existing file: " + path, e);
         }
+
+        // Copy the file to the relevant directory
         try {
-            if (file.isEmpty()) {
-                EventLogger.logger.info("File is empty: {}", path);
-                throw new StorageException("File is empty: " + path);
-            }
-            logger.debug("Copying file {} to {}", file.getOriginalFilename(), this.rootLocation);
-            EventLogger.logger.info("Copying file {} to {}", file.getOriginalFilename(), this.rootLocation);
+            logger.debug("Copying file {} to {}", file.getOriginalFilename(), path);
+            EventLogger.logger.info("Copying file {} to {}", file.getOriginalFilename(), path);
             Files.copy(file.getInputStream(), path);
         } catch (IOException e) {
-            EventLogger.logger.info("Failed to store file in shared directory {}", this.rootLocation);
-            throw new StorageException("Failed to store file in shared directory" + this.rootLocation, e);
+            EventLogger.logger.info("Failed to store file in shared directory {}", path);
+            throw new StorageException("Failed to store file in shared directory " + path, e);
         }
     }
